no explicit weight casting.

+++ examples/text_to_image/train_text_to_image_lora.py
@@ -618,9 +618,12 @@ def main():
     # Move text_encoder and vae to gpu and cast to weight_dtype
     # The unet's weight_dtype is kept in full precision because of autocasts
     # during the backward pass.
-    unet.to(accelerator.device, dtype=weight_dtype)
-    text_encoder.to(accelerator.device, dtype=weight_dtype)
-    vae.to(accelerator.device, dtype=weight_dtype)
+    # unet.to(accelerator.device, dtype=weight_dtype)
+    # text_encoder.to(accelerator.device, dtype=weight_dtype)
+    # vae.to(accelerator.device, dtype=weight_dtype)
+    unet.to(accelerator.device)
+    text_encoder.to(accelerator.device)
+    vae.to(accelerator.device)
 
     # We need to recalculate our total training steps as the size of the training dataloader may have changed.
     num_update_steps_per_epoch = math.ceil(len(train_dataloader) / args.gradient_accumulation_steps)
